requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chain==1.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.1.0
faiss-cpu==1.8.0.post1
fastjsonschema==2.21.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
idna==3.8
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lang==0.1
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.15
langchain-openai==0.2.4
langchain-text-splitters==0.3.0
langgraph==0.2.43
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.128
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.7.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
numpy==1.26.4
openai==1.52.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
streamlit==1.38.0
SudachiDict-core==20240716
SudachiPy==0.6.8
tenacity==8.5.0
tiktoken==0.7.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.11.1
